Decouple docker logic from API endpoints
