+++ packages/dappmanager/src/calls/packageRestart.ts
@@ -1,10 +1,5 @@
-import fs from "fs";
-import params from "../params";
-import * as getPath from "../utils/getPath";
-import { restartDappmanagerPatch } from "../modules/installer/restartPatch";
-import { dockerComposeRm } from "../modules/docker/dockerCommands";
-import { dockerComposeUpSafe } from "../modules/docker/dockerSafe";
 import * as eventBus from "../eventBus";
+import { restartPackage } from "../modules/docker/restartPackage";
 
 /**
  * Calls docker rm and docker up on a package
@@ -14,20 +9,9 @@ import * as eventBus from "../eventBus";
 export async function packageRestart({ id }: { id: string }): Promise<void> {
   if (!id) throw Error("kwarg id must be defined");
 
-  const composePath = getPath.dockerComposeSmart(id);
-  if (!fs.existsSync(composePath)) {
-    throw Error(`No docker-compose found: ${composePath}`);
-  }
+  await restartPackage(id);
 
-  if (id.includes(params.dappmanagerDnpName)) {
-    await restartDappmanagerPatch({ composePath });
-  } else {
-    // Combining rm && up doesn't prevent the installer from crashing
-    await dockerComposeRm(composePath);
-    await dockerComposeUpSafe(composePath);
-
-    // Emit packages update
-    eventBus.requestPackages.emit();
-    eventBus.packagesModified.emit({ ids: [id] });
-  }
+  // Emit packages update
+  eventBus.requestPackages.emit();
+  eventBus.packagesModified.emit({ ids: [id] });
 }
